requirements-dev.txt
deepdiff==3.3.0
django-debug-toolbar==1.10.1
django-extensions==2.1.3
jsonpickle==1.0
mypy==0.630
mypy-extensions==0.4.1
six==1.11.0
sqlparse==0.2.4
typed-ast==1.1.0
